+++ src/main/java/org/folio/rest/impl/EventConfigAPIs.java
@@ -19,6 +19,7 @@ import javax.ws.rs.core.Response;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Optional;
 
 import static javax.ws.rs.core.MediaType.TEXT_HTML;
 import static javax.ws.rs.core.MediaType.TEXT_PLAIN;
@@ -28,6 +29,7 @@ public class EventConfigAPIs implements EventConfig {
 
   private static final String ERROR_RUNNING_VERTICLE = "Error running on verticle for `%s`: %s";
   private static final String ERROR_EVENT_CONFIG_NOT_FOUND = "Event Config with ID: `%s` was not found in the db";
+  private static final String ERROR_EVENT_CONFIG_NOT_FOUND_NAME = "Event Config with `name`: `%s` was not found in the db";
 
   private final Logger logger = LoggerFactory.getLogger(EventConfigAPIs.class);
 
@@ -56,11 +58,12 @@ public class EventConfigAPIs implements EventConfig {
     try {
       List<Template> templates = entity.getTemplates();
       templates.forEach(template -> template.setOutputFormat(findOutputFormat(template)));
+      changeNameToUpperCase(entity);
       JsonObject entityJson = JsonObject.mapFrom(entity);
       context.runOnContext(contextHandler -> service.createEventConfig(tenantId, entityJson, serviceHandler -> {
         if (serviceHandler.failed()) {
           String errorMessage = serviceHandler.cause().getMessage();
-          asyncHandler.handle(createFutureResponse(PostEventConfigResponse.respond500WithTextPlain(errorMessage)));
+          asyncHandler.handle(createFutureResponse(PostEventConfigResponse.respond400WithTextPlain(errorMessage)));
           return;
         }
 
@@ -119,7 +122,9 @@ public class EventConfigAPIs implements EventConfig {
           asyncHandler.handle(createFutureResponse(GetEventConfigByIdResponse.respond500WithTextPlain(errorMessage)));
           return;
         }
-        if (Objects.isNull(serviceHandler.result())) {
+        JsonObject jsonObject = serviceHandler.result();
+        Boolean isNotFound = Optional.ofNullable(jsonObject.getBoolean(VALUE_IS_NOT_FOUND)).orElse(false);
+        if (isNotFound) {
           String message = String.format(ERROR_EVENT_CONFIG_NOT_FOUND, id);
           logger.debug(message);
           EventResponse eventResponse = new EventResponse().withMessage(message);
@@ -150,6 +155,7 @@ public class EventConfigAPIs implements EventConfig {
   public void putEventConfigById(String id, EventEntity entity, Map<String, String> headers,
                                  Handler<AsyncResult<Response>> asyncHandler, Context context) {
     try {
+      changeNameToUpperCase(entity);
       final JsonObject entityJson = JsonObject.mapFrom(entity);
       context.runOnContext(contextHandler -> service.updateEventConfig(tenantId, id, entityJson, serviceHandler -> {
         if (serviceHandler.failed()) {
@@ -217,6 +223,49 @@ public class EventConfigAPIs implements EventConfig {
     }
   }
 
+  /**
+   * Get event config by name
+   *
+   * @param name         the configuration event name
+   * @param asyncHandler an AsyncResult<Response> Handler {@link Handler}
+   *                     which must be called as follows in the final callback (most internal callback) of the function
+   */
+  @Override
+  public void getEventConfigNameByName(String name, Map<String, String> okapiHeaders,
+                                       Handler<AsyncResult<Response>> asyncHandler, Context context) {
+    try {
+      String configName = name.trim().toUpperCase();
+      context.runOnContext(contextHandler -> service.findEventConfigByName(tenantId, configName,
+        serviceHandler -> {
+          if (serviceHandler.failed()) {
+            String errorMessage = serviceHandler.cause().getMessage();
+            asyncHandler.handle(createFutureResponse(GetEventConfigNameByNameResponse
+              .respond500WithTextPlain(errorMessage)));
+            return;
+          }
+          JsonObject jsonObject = serviceHandler.result();
+          Boolean isNotFound = Optional.ofNullable(jsonObject.getBoolean(VALUE_IS_NOT_FOUND)).orElse(false);
+          if (isNotFound) {
+            String message = String.format(ERROR_EVENT_CONFIG_NOT_FOUND_NAME, name);
+            logger.debug(message);
+            EventResponse eventResponse = new EventResponse().withMessage(message);
+            asyncHandler.handle(createFutureResponse(
+              GetEventConfigNameByNameResponse.respond404WithApplicationJson(eventResponse)));
+            return;
+          }
+
+          EventEntity eventEntity = getResponseEntity(serviceHandler, EventEntity.class);
+          asyncHandler.handle(createFutureResponse(GetEventConfigNameByNameResponse
+            .respond200WithApplicationJson(eventEntity)));
+        }));
+    } catch (Exception ex) {
+      String errorMessage = String.format(ERROR_RUNNING_VERTICLE, "getEventConfigNameByName", ex.getMessage());
+      logger.error(errorMessage, ex);
+      asyncHandler.handle(createFutureResponse(GetEventConfigNameByNameResponse
+        .respond500WithTextPlain(errorMessage)));
+    }
+  }
+
   private String findOutputFormat(Template template) {
     String outputFormat = template.getOutputFormat();
     if (StringUtils.isBlank(outputFormat)) {
@@ -226,5 +275,10 @@ public class EventConfigAPIs implements EventConfig {
       return TEXT_HTML;
     }
     return TEXT_PLAIN;
+  }
+
+  private void changeNameToUpperCase(EventEntity entity) {
+    String configName = entity.getName().trim().toUpperCase();
+    entity.setName(configName);
   }
 }