+++ src/test/java/org/folio/impl/EventConfigAPIsTest.java
@@ -36,6 +36,7 @@ import java.util.Objects;
 import java.util.UUID;
 
 import static junit.framework.TestCase.assertEquals;
+import static junit.framework.TestCase.assertTrue;
 import static org.folio.rest.RestVerticle.OKAPI_HEADER_TENANT;
 import static org.folio.rest.RestVerticle.OKAPI_HEADER_TOKEN;
 
@@ -55,6 +56,7 @@ public class EventConfigAPIsTest {
   private static final Logger logger = LoggerFactory.getLogger(EventConfigAPIsTest.class);
 
   private static String restPath;
+  private static String restPathGetByName;
   private static RequestSpecification request;
   private static Vertx vertx;
   private static String useExternalDatabase;
@@ -71,6 +73,7 @@ public class EventConfigAPIsTest {
       new Header(OKAPI_HEADER_TENANT, TENANT_ID),
       new Header(OKAPI_HEADER_TOKEN, OKAPI_TOKEN_VAL));
     restPath = System.getProperty("org.folio.event.config.rest.path", "/eventConfig");
+    restPathGetByName = System.getProperty("org.folio.event.config.rest.path", "/eventConfig/name");
     useExternalDatabase = System.getProperty("org.folio.event.config.test.database", EXTERNAL_DATABASE_VAL);
     request = RestAssured.given()
       .port(port)
@@ -120,20 +123,23 @@ public class EventConfigAPIsTest {
     }));
   }
 
-
   @Before
   public void setUp(TestContext context) {
+    Async async = context.async();
     PostgresClient.getInstance(vertx, TENANT_ID).delete(SNAPSHOTS_TABLE_NAME, new Criterion(), event -> {
       if (event.failed()) {
         logger.error(event.cause());
         context.fail(event.cause());
+      } else {
+        async.complete();
       }
     });
   }
 
   @Test
   public void testFullRestAPI() {
-    JsonObject expectedEntity = getJsonEntity(UUID.randomUUID().toString(), "test", true, new JsonArray());
+    JsonObject expectedEntity = getJsonEntity(UUID.randomUUID().toString(), "reset_password", true, new JsonArray());
+    expectedEntity.put("name", "RESET_PASSWORD");
 
     // create a new event config
     Response response = requestPostEventConfig(expectedEntity)
@@ -167,7 +173,7 @@ public class EventConfigAPIsTest {
   @Test
   public void testPostEventConfig() {
     String id = UUID.randomUUID().toString();
-    JsonObject expectedEntity = getJsonEntity(id, "test name", false, new JsonArray());
+    JsonObject expectedEntity = getJsonEntity(id, "RESET_PASSWORD", false, new JsonArray());
     String expectedDeleteMessage = "Configuration with id: '%s' was successfully deleted";
 
     // create a new event config
@@ -194,7 +200,7 @@ public class EventConfigAPIsTest {
   @Test
   public void testPostHtmlEventConfig() {
     JsonArray templates = createTemplates("email", "text/html");
-    JsonObject expectedEntity = getJsonEntity(UUID.randomUUID().toString(), "test", true, templates);
+    JsonObject expectedEntity = getJsonEntity(UUID.randomUUID().toString(), "RESET_PASSWORD", true, templates);
 
     // create a new event config
     Response response = requestPostEventConfig(expectedEntity)
@@ -244,11 +250,76 @@ public class EventConfigAPIsTest {
   }
 
   @Test
-  public void testGetAllEventEntries() {
-    requestPostEventConfig(getJsonEntity(null, RandomStringUtils.randomAlphabetic(10), false, new JsonArray()))
+  public void testUniqueNameOfConfiguration() {
+    String configName = "RESET_PASSWORD";
+    requestPostEventConfig(getJsonEntity(UUID.randomUUID().toString(), configName,
+      false, new JsonArray()))
       .then()
       .statusCode(HttpStatus.SC_CREATED);
-    requestPostEventConfig(getJsonEntity(null, RandomStringUtils.randomAlphabetic(20), false, new JsonArray()))
+
+    Response response = requestPostEventConfig(getJsonEntity(UUID.randomUUID().toString(), configName,
+      false, new JsonArray()))
+      .then()
+      .statusCode(HttpStatus.SC_BAD_REQUEST)
+      .extract()
+      .response();
+
+    String body = response.getBody().print();
+    assertTrue(body.contains("duplicate key value violates unique constraint"));
+  }
+
+  @Test
+  public void testGetEventConfigByName() {
+    String configName = "CREATE_PASSWORD";
+    JsonObject expectedEntity = getJsonEntity(UUID.randomUUID().toString(), configName, true, new JsonArray());
+
+    String responseBody = requestPostEventConfig(expectedEntity)
+      .then()
+      .statusCode(HttpStatus.SC_CREATED)
+      .extract()
+      .response()
+      .getBody().print();
+    String actualId = new JsonObject(responseBody).getString("id");
+    expectedEntity.put("id", actualId);
+
+    Response response = requestGetEventByName(configName)
+      .then()
+      .statusCode(HttpStatus.SC_OK)
+      .extract()
+      .response();
+
+    assertEquals(expectedEntity, new JsonObject(response.getBody().print()));
+  }
+
+  @Test
+  public void testGetEventConfigByIncorrectName() {
+    String configName = "password";
+    JsonObject expectedEntity = getJsonEntity(UUID.randomUUID().toString(), configName, true, new JsonArray());
+    requestPostEventConfig(expectedEntity)
+      .then()
+      .statusCode(HttpStatus.SC_CREATED)
+      .extract()
+      .response();
+
+    Response response = requestGetEventByName("INCORRECT")
+      .then()
+      .statusCode(HttpStatus.SC_NOT_FOUND)
+      .extract()
+      .response();
+
+    String body = response.getBody().print();
+    String message = new JsonObject(body).getString("message");
+    assertTrue(message.contains("Event Config with `name`: `INCORRECT` was not found in the db"));
+  }
+
+  @Test
+  public void testGetAllEventEntries() {
+    requestPostEventConfig(getJsonEntity(null, RandomStringUtils.randomAlphabetic(10),
+      false, new JsonArray()))
+      .then()
+      .statusCode(HttpStatus.SC_CREATED);
+    requestPostEventConfig(getJsonEntity(null, RandomStringUtils.randomAlphabetic(20),
+      false, new JsonArray()))
       .then()
       .statusCode(HttpStatus.SC_CREATED);
 
@@ -279,7 +350,8 @@ public class EventConfigAPIsTest {
   @Test
   public void testPutEventConfigById() {
     String id = UUID.randomUUID().toString();
-    JsonObject entity = getJsonEntity("0002", RandomStringUtils.randomAlphabetic(20), true, new JsonArray());
+    JsonObject entity = getJsonEntity("0002", RandomStringUtils.randomAlphabetic(20),
+      true, new JsonArray());
 
     Response response = requestPutEventConfig(id, entity)
       .then()
@@ -325,6 +397,12 @@ public class EventConfigAPIsTest {
     return request
       .when()
       .get(String.format(PATH_TEMPLATE, restPath, expectedId));
+  }
+
+  private Response requestGetEventByName(String name) {
+    return request
+      .when()
+      .get(String.format(PATH_TEMPLATE, restPathGetByName, name));
   }
 
   private Response requestGetEventEntries() {